cmake: Move parsing the CMake trace into the CMake module
