+++ mesonbuild/cmake/__init__.py
@@ -20,9 +20,13 @@ __all__ = [
     'CMakeExecutor',
     'CMakeException',
     'CMakeInterpreter',
+    'CMakeTarget',
+    'CMakeTraceLine',
+    'CMakeTraceParser',
 ]
 
 from .common import CMakeException
 from .client import CMakeClient
 from .executor import CMakeExecutor
 from .interpreter import CMakeInterpreter
+from .traceparser import CMakeTarget, CMakeTraceLine, CMakeTraceParser
